% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_linear.R
\name{kernel_linear}
\alias{kernel_linear}
\title{Variance Component Linear Kernel Matrix}
\usage{
kernel_linear(dose, method = "linear")
}
\arguments{
\item{dose}{data.frame or matrix with}

\item{method}{type of kernel; currently only linear kernel implemented}
}
\value{
square symmetric kernel matrix for subject similarity by genotype dosage
}
\description{
Variance component Linear kernel matrix from genotype dosage
}
\examples{
data(vcexample)
Kern1 <- kernel_linear(dose[,which(doseinfo[,1]==1)], method="linear")
Kern1[1:5,1:5]

}
\seealso{
\code{\link{vcpen}}
}
\author{
JP Sinnwell, DJ Schaid
}
